Fix issue where NIF can be found when erlaudio is a dep
Prior to this commit, when erlaudio is a dependency of an elixir
application, the erlaudio_drv NIF can't be located because
erlaudio_drv.so is in the priv_dir for erlaudio, not erlaudio_drv. This
commit explicitly looks for the NIF in the priv dir for the erlaudio
module rather than the ?MODULE macro which expands to erlaudio_drv in
this case.